+++ 1.21.11/common/src/main/java/com/iamkaf/liteminer/shapes/StaircaseDownWalker.java
@@ -11,11 +11,8 @@ import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 
 import java.util.HashSet;
-import java.util.Set;
 
 public class StaircaseDownWalker implements Walker {
-    public final Set<BlockPos> VISITED = new HashSet<>();
-
     @Override
     public String toString() {
         return "Staircase Down";
@@ -37,16 +34,16 @@ public class StaircaseDownWalker implements Walker {
             return potentialBrokenBlocks;
         }
 
-        searchBlocks(player, level, origin, origin, potentialBrokenBlocks, originState.getBlock(), direction);
-        VISITED.clear();
+        HashSet<BlockPos> visited = new HashSet<>();
+        searchBlocks(player, level, origin, origin, potentialBrokenBlocks, originState.getBlock(), direction, visited);
 
         return potentialBrokenBlocks;
     }
 
     private void searchBlocks(Player player, Level level, BlockPos myPos, BlockPos absoluteOrigin,
-            HashSet<BlockPos> blocksToCollapse, Block originBlock, Direction direction) {
-        if (VISITED.size() >= Liteminer.CONFIG.blockBreakLimit.get()) return;
-        if (VISITED.contains(myPos)) return;
+            HashSet<BlockPos> blocksToCollapse, Block originBlock, Direction direction, HashSet<BlockPos> visited) {
+        if (visited.size() >= Liteminer.CONFIG.blockBreakLimit.get()) return;
+        if (visited.contains(myPos)) return;
 
         BlockState state = level.getBlockState(myPos);